requirements.txt
streamlit
google-generativeai
chromadb
pypdf